+++ src/main/java/com/braintreegateway/SandboxValues.java
@@ -55,7 +55,7 @@ public class SandboxValues {
     }
 
     public enum TransactionAmount {
-        AUTHORIZE("1000.00"), DECLINE("2000.00"), FAILED("3000.00");
+        AUTHORIZE("1000.00"), DECLINE("2000.00"), HARD_DECLINE("2015.00"), FAILED("3000.00");
 
         public BigDecimal amount;